Adding parameter to select fields in a query
The firestore api allows to only get certain fields in a query via a 'select' parameter. This can be used to reduce the size of the returned messages. This commit adds the select parameter to the query command by which several fieldpathes can be given.

Example: fuego query Collection --select FieldA --select FieldB